Move ResultSet -> Seq[InternalRow] conversion into JdbcUtils
